#4 Creating  generic repository and service and controller

+++ EducationalInstitution/EducationalInstitution.Api/Models/Output/RegistrationInvoiceOutput.cs
@@ -7,7 +7,7 @@ namespace EducationalInstitution.Api.Models.Output
         public DateTimeOffset CreationDate { get; set; }
         public int Id { get; set; }
         public decimal TotalTuition { get; set; }
-        public decimal TotalNumberCourses { get; set; }
+        public int TotalNumberCourses { get; set; }
         public string TrackingCode { get; set; }
         public int DepositID { get; set; }
         public string Description { get; set; }
